feat: include embedded knowledge file in zip package

+++ packages/fx-core/tests/component/driver/teamsApp/createAppPackage.test.ts
@@ -13,6 +13,7 @@ import { MockedLogProvider, MockedUserInteraction } from "../../../plugins/solut
 import { FileNotFoundError, JSONSyntaxError } from "../../../../src/error/common";
 import { manifestUtils } from "../../../../src/component/driver/teamsApp/utils/ManifestUtils";
 import {
+  DeclarativeCopilotManifestSchema,
   err,
   ok,
   Platform,
@@ -23,6 +24,8 @@ import {
 import AdmZip from "adm-zip";
 import { InvalidFileOutsideOfTheDirectotryError } from "../../../../src/error/teamsApp";
 import { MockedM365Provider } from "../../../core/utils";
+import { copilotGptManifestUtils } from "../../../../src/component/driver/teamsApp/utils/CopilotGptManifestUtils";
+import { FeatureFlags, featureFlagManager } from "../../../../src/common/featureFlags";
 
 describe("teamsApp/createAppPackage", async () => {
   const teamsAppDriver = new CreateAppPackageDriver();
@@ -1733,6 +1736,83 @@ describe("teamsApp/createAppPackage", async () => {
             openapiContent.search(fakeUrl) >= 0 &&
             openapiContent.search(openapiServerPlaceholder) < 0
         );
+        await fs.remove(args.outputZipPath);
+      }
+    });
+
+    it("should add embedded knowledge files for Copilot GPT", async () => {
+      // Enable builder API flag.
+      sinon
+        .stub(featureFlagManager, "getBooleanValue")
+        .withArgs(FeatureFlags.BuilderAPIEnabled)
+        .returns(true);
+
+      const args: CreateAppPackageArgs = {
+        manifestPath:
+          "./tests/plugins/resource/appstudio/resources-multi-env/templates/appPackage/v3.manifest.template.json",
+        outputZipPath:
+          "./tests/plugins/resource/appstudio/resources-multi-env/build/appPackage/appPackage.embedded.zip",
+        outputJsonPath:
+          "./tests/plugins/resource/appstudio/resources-multi-env/build/appPackage/manifest.embedded.json",
+      };
+
+      const manifest = new TeamsAppManifest();
+      manifest.copilotAgents = {
+        declarativeAgents: [{ file: "resources/declarativeAgent.json", id: "1" }],
+      };
+      manifest.icons = {
+        color: "resources/color.png",
+        outline: "resources/outline.png",
+      };
+
+      // Updated gpt manifest stub with required properties.
+      const declarativeAgentManifest = {
+        name: "TestDeclarativeCopilot",
+        description: "Test declarative copilot manifest",
+        actions: [],
+        capabilities: [
+          {
+            name: "EmbeddedKnowledge",
+            files: [{ file: "knowledge.docx" }],
+          },
+        ],
+      } as DeclarativeCopilotManifestSchema;
+
+      sinon.stub(manifestUtils, "getManifestV3").resolves(ok(manifest));
+      sinon.stub(fs, "chmod").callsFake(async () => {});
+      sinon.stub(fs, "writeFile").callsFake(async () => {});
+      sinon.stub(copilotGptManifestUtils, "getManifest").resolves(ok(declarativeAgentManifest));
+      sinon.stub(fs, "pathExists").callsFake(async (filePath) => {
+        // Return true for all required files including declarativeAgent.json, color/outline files and knowledge file.
+        if (
+          filePath.includes("knowledge.docx") ||
+          filePath.includes("declarativeAgent.json") ||
+          filePath.includes("color.png") ||
+          filePath.includes("outline.png")
+        ) {
+          return true;
+        }
+        return true;
+      });
+      sinon.stub(fs, "stat").resolves({ mode: 0o777 } as any);
+
+      const mockedDriverContext: any = {
+        m365TokenProvider: {},
+        projectPath: "./",
+        platform: 0,
+        logProvider: { info: () => {} },
+        ui: {},
+        addTelemetryProperties: () => {},
+      };
+
+      const result = (await teamsAppDriver.execute(args, mockedDriverContext)).result;
+      chai.assert.isTrue(result.isOk());
+
+      if (await fs.pathExists(args.outputZipPath)) {
+        const AdmZip = require("adm-zip");
+        const zip = new AdmZip(args.outputZipPath);
+        const knowledgeEntry = zip.getEntry("knowledge.docx");
+        chai.assert.exists(knowledgeEntry, "Embedded knowledge file should be added");
         await fs.remove(args.outputZipPath);
       }
     });
